refactor(stores): complete ADR-048 core→infrastructure migration (#935)
Move message_index, prefs_store, thread_store, pairing from
core/stores/ to infrastructure/stores/. Update all import sites
(src + tests + TYPE_CHECKING in core/hub). Remove 6 stale
import-linter ignore entries, add entries for new TYPE_CHECKING
paths. Add folder size exemption for infrastructure/stores/.
