+++ .importlinter
@@ -13,14 +13,8 @@ layers =
     lyra.llm | lyra.nats
     lyra.core
 ignore_imports =
-    # Transitional: core.stores files pending migration import infrastructure (ADR-048)
+    # Transitional: agent_store_protocol remains in core (Direction 2, out of scope)
     lyra.core.stores.agent_store_protocol -> lyra.infrastructure.stores.agent_store
-    lyra.core.stores.pairing -> lyra.infrastructure.stores.sqlite_base
-    lyra.core.stores.pairing -> lyra.infrastructure.stores.auth_store
-    lyra.core.stores.message_index -> lyra.infrastructure.stores.sqlite_base
-    lyra.core.stores.thread_store -> lyra.infrastructure.stores.sqlite_base
-    lyra.core.stores.prefs_store -> lyra.infrastructure.stores.sqlite_base
-    lyra.core.stores.prefs_store -> lyra.infrastructure.stores.identity_alias_store
     # Downward TYPE_CHECKING imports (ADR-048 transition)
     lyra.core.agent.agent_refiner -> lyra.infrastructure.stores.agent_store
     lyra.core.auth.authenticator -> lyra.infrastructure.stores.auth_store
@@ -29,6 +23,13 @@ ignore_imports =
     lyra.core.memory.memory -> lyra.infrastructure.stores.identity_alias_store
     lyra.core.hub.hub -> lyra.infrastructure.stores.identity_alias_store
     lyra.core.hub.hub_registration -> lyra.infrastructure.stores.identity_alias_store
+    # TYPE_CHECKING imports for migrated stores (#935)
+    lyra.core.hub.hub -> lyra.infrastructure.stores.message_index
+    lyra.core.hub.hub -> lyra.infrastructure.stores.pairing
+    lyra.core.hub.hub -> lyra.infrastructure.stores.prefs_store
+    lyra.core.hub.hub_registration -> lyra.infrastructure.stores.message_index
+    lyra.core.hub.hub_shutdown -> lyra.infrastructure.stores.message_index
+    lyra.core.pool.pool_observer -> lyra.infrastructure.stores.message_index
     # TurnStore moved to infrastructure per ADR-048 (#760 review fix f2)
     lyra.core.hub.hub -> lyra.infrastructure.stores.turn_store
     lyra.core.hub.hub_registration -> lyra.infrastructure.stores.turn_store
