+++ src/lyra/adapters/discord/discord_threads.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from lyra.core.messaging.message import InboundMessage
-    from lyra.core.stores.thread_store import ThreadStore
+    from lyra.infrastructure.stores.thread_store import ThreadStore
 
 log = logging.getLogger(__name__)
