+++ src/lyra/bootstrap/factory/unified.py
@@ -40,7 +40,7 @@ from lyra.core.config import HubConfig
 from lyra.core.hub import Hub
 from lyra.core.hub.event_bus import PipelineEventBus
 from lyra.core.messaging.message import InboundMessage
-from lyra.core.stores.pairing import PairingManager, set_pairing_manager
+from lyra.infrastructure.stores.pairing import PairingManager, set_pairing_manager
 from lyra.nats.nats_bus import NatsBus
 from lyra.nats.queue_groups import HUB_INBOUND
