+++ src/lyra/infrastructure/stores/__init__.py
@@ -9,7 +9,15 @@ from lyra.infrastructure.stores.credential_store import (
     LyraKeyring,
 )
 from lyra.infrastructure.stores.identity_alias_store import IdentityAliasStore
+from lyra.infrastructure.stores.message_index import MessageIndex
+from lyra.infrastructure.stores.pairing import (
+    PairingManager,
+    get_pairing_manager,
+    set_pairing_manager,
+)
+from lyra.infrastructure.stores.prefs_store import PrefsStore, UserPrefs
 from lyra.infrastructure.stores.sqlite_base import SqliteStore
+from lyra.infrastructure.stores.thread_store import ThreadStore
 from lyra.infrastructure.stores.turn_store import TurnStore
 
 __all__ = [
@@ -20,6 +28,13 @@ __all__ = [
     "CredentialStore",
     "IdentityAliasStore",
     "LyraKeyring",
+    "MessageIndex",
+    "PairingManager",
+    "PrefsStore",
     "SqliteStore",
+    "ThreadStore",
     "TurnStore",
+    "UserPrefs",
+    "get_pairing_manager",
+    "set_pairing_manager",
 ]
