+++ src/lyra/infrastructure/stores/message_index.py
