+++ tests/core/conftest.py
@@ -36,9 +36,9 @@ from lyra.core.messaging.message import (
 )
 from lyra.core.messaging.render_events import RenderEvent
 from lyra.core.pool import Pool
-from lyra.core.stores.pairing import PairingConfig, PairingManager
 from lyra.infrastructure.stores.agent_store import AgentRow, AgentStore
 from lyra.infrastructure.stores.auth_store import AuthStore
+from lyra.infrastructure.stores.pairing import PairingConfig, PairingManager
 
 # ---------------------------------------------------------------------------
 # MessageManager shared constants
@@ -803,7 +803,7 @@ _open_pairing_stores: list[AuthStore] = []
 @pytest.fixture(autouse=True)
 async def _cleanup_pairing_state(tmp_path: Path):
     """Reset pairing global and close all PairingManagers/AuthStores after each test."""
-    from lyra.core.stores.pairing import set_pairing_manager
+    from lyra.infrastructure.stores.pairing import set_pairing_manager
 
     setattr(_cleanup_pairing_state, "tmp_path", tmp_path)
     yield
