+++ tests/core/test_pairing_commands.py
@@ -10,7 +10,7 @@ from unittest.mock import MagicMock
 from lyra.commands.pairing.handlers import cmd_invite, cmd_join, cmd_unpair
 from lyra.core.auth.trust import TrustLevel
 from lyra.core.pool import Pool
-from lyra.core.stores.pairing import set_pairing_manager
+from lyra.infrastructure.stores.pairing import set_pairing_manager
 
 from .conftest import (
     _PAIRING_ADMIN_ID as _ADMIN_ID,
