+++ tests/core/test_prefs_alias.py
@@ -6,8 +6,8 @@ from pathlib import Path
 
 import pytest
 
-from lyra.core.stores.prefs_store import PrefsStore
 from lyra.infrastructure.stores.identity_alias_store import IdentityAliasStore
+from lyra.infrastructure.stores.prefs_store import PrefsStore
 
 # ---------------------------------------------------------------------------
 # Fixtures
