+++ tests/core/test_prefs_store.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-from lyra.core.stores.prefs_store import PrefsStore, UserPrefs
+from lyra.infrastructure.stores.prefs_store import PrefsStore, UserPrefs
 
 # ---------------------------------------------------------------------------
 # Fixtures
