+++ tools/folder_exemptions.txt
@@ -7,3 +7,4 @@
 # V5 completed under #848 — core/hub/ decomposed into middleware/, outbound/, pipeline/.
 
 src/lyra/core           # 13 files — #858 config dataclass extraction
+src/lyra/infrastructure/stores  # 14 files — #935 ADR-048 store migration (4 files moved from core/stores)
